IConverionForm interface moved to interfaces.py
